Move querying patients to App.js
