+++ speedy-recovery-client/src/components/secured/patientinformation/PractitionerInfo.js
@@ -21,36 +21,31 @@ about each of the practitioners' patients.
 import React, { Component } from "react";
 import {Link} from "react-router-dom";
 import { List,Image } from "semantic-ui-react";
-import FhirDataQueryingService from "../../../service/FhirDataQueryingService";
-import {fhirMapPerson} from "../../../service/FhirDataMappingService";
 
 class PractitionerInfo extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            userList: [],
             patients: []
         };
     }
 
     render() {
-            const patientItems = this.state.patients.map(patient => {
-                return (
-                    <List.Item as={Link} to={{
-                        pathname: `/secured/information/${patient.id}`,
-                        state: {
-                            id: patient.id,
-                            name: patient.name,
-                            patients : this.state.patients,
-                            fhirVersion : this.props.fhirVersion
-                            }}}
-                               key={patient.id}>
-                        <Image avatar src='https://institutogoldenprana.com.br/wp-content/uploads/2015/08/no-avatar-25359d55aa3c93ab3466622fd2ce712d1.jpg' />
-                        <List.Content>
-                            <List.Header>{patient.name}</List.Header>
-                        </List.Content>
-                    </List.Item>
-                );});
+        const patientItems = this.state.patients.map(patient => {
+            return (
+                <List.Item as={Link} to={{
+                    pathname: `/secured/information/${patient.id}`,
+                    state: {
+                        patient : patient,
+                        fhirVersion : this.props.fhirVersion
+                    }}}
+                           key={patient.id}>
+                    <Image avatar src='https://institutogoldenprana.com.br/wp-content/uploads/2015/08/no-avatar-25359d55aa3c93ab3466622fd2ce712d1.jpg' />
+                    <List.Content>
+                        <List.Header>{patient.name}</List.Header>
+                    </List.Content>
+                </List.Item>
+            );});
         return (
             <div>
                 <h1>My patients </h1>
@@ -61,26 +56,9 @@ class PractitionerInfo extends Component {
         );
     }
     componentWillMount() {
-        if(this.props.userList){
-            this.updateStatePatient(this.props.userList);
+        if(this.props.patients){
+            this.setState({patients : this.props.patients})
         }
-    }
-
-    updateStatePatient(userList) {
-        const patients = [];
-        for(let user of userList){
-            FhirDataQueryingService.getPatient(user.id)
-                .then(patientResource => {
-                    const patient = patientResource.map(patient => fhirMapPerson(patient, this.props.fhirVersion)
-                    );
-                    patients.push(patient[0]);
-                    this.setState({patients})
-                })
-                .catch(error => {
-                    console.error(error);
-                });
-        }
-
     }
 }
